run: stop propagating opc through vararg instructions
Reasoning for correctness:

The only instructions that actually care about the number of parameters are
EQUAL? and CALL. EQUAL? was dealt with in a previous commit. The reasoning for
CALL is as follows:

The only way for CALL to be called with an "inherited" opc is if no parameters
are passed. In that case, all parameters will be undefined. There is no
difference if they are omitted from the parameters, as assigning a parameter to
be "undefined" actually makes it 0 (because the call-stack entry is stored in a
Uint16Array, and undefined is implicitly coerced to 0). Since each entry is only
assigned once, and the default value is already 0, not performing the assignment
has no observable effect.